pyproject.toml
[project]
name = "microreview"
version = "0.1.0"
description = "A modular, explainable AI code review GitHub App powered by DSPy micro-agents"
authors = [
    { name = "MicroReview Contributors" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
keywords = ["ai", "code-review", "github", "dspy", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    # Core framework for LLM-powered micro-agents (README requirement)
    "dspy-ai>=2.4.0",
    
    # YAML configuration file parsing (.microreview.yml files)
    "pyyaml>=6.0",
    
    # GitHub API integration (for production GitHub App deployment)
    "PyGithub>=1.59.0",
    
    # Web framework for webhook handling (for GitHub App)
    "flask>=2.3.0",
    
    # LLM providers (README: Multi-Provider Support)
    "openai>=1.0.0",      # For OpenAI GPT models (README example: gpt-4o)
    "anthropic>=0.7.0",   # For Anthropic Claude models (README example: claude-3-sonnet)
]

[project.optional-dependencies]
dev = [
    # Development and testing dependencies
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",      # Code formatting
    "flake8>=6.0.0",      # Linting
]

[project.urls]
Homepage = "https://github.com/AnthonyHerman/MicroReview"
Repository = "https://github.com/AnthonyHerman/MicroReview"
Issues = "https://github.com/AnthonyHerman/MicroReview/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "config", "core", "tests", "docs"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
